+++ packages/react/src/components/Menu/Menu.tsx
@@ -32,7 +32,7 @@ export interface MenuProps extends UIComponentProps, ChildrenComponentProps {
   /**
    * Accessibility behavior if overridden by the user.
    * @default menuBehavior
-   * @available toolbarBehavior, tabListBehavior
+   * @available menuAsToolbarBehavior, tabListBehavior
    * */
   accessibility?: Accessibility