+++ packages/react/src/components/Menu/MenuItem.tsx
@@ -46,7 +46,7 @@ export interface MenuItemProps
   /**
    * Accessibility behavior if overridden by the user.
    * @default menuItemBehavior
-   * @available toolbarButtonBehavior, tabBehavior
+   * @available menuItemAsToolbarButtonBehavior, tabBehavior
    * */
   accessibility?: Accessibility